+++ src/main/java/org/bytedeco/javacpp/tools/ClassScanner.java
@@ -91,7 +91,7 @@ class ClassScanner {
 
     public void addPackage(String packageName, boolean recursive) throws IOException, ClassNotFoundException, NoClassDefFoundError {
         String[] paths = loader.getPaths();
-        final String packagePath = packageName == null ? null : (packageName.replace('.', '/') + "/");
+        final String packagePath = packageName != null && packageName.length() > 0 ? (packageName.replace('.', '/') + "/") : packageName;
         int prevSize = classes.size();
         for (String p : paths) {
             File file = new File(p);
@@ -108,9 +108,8 @@ class ClassScanner {
                 jis.close();
             }
         }
-        if (classes.size() == 0 && packageName == null) {
+        if (classes.size() == 0 && (packageName == null || packageName.length() == 0)) {
             logger.warn("No classes found in the unnamed package");
-            Builder.printHelp();
         } else if (prevSize == classes.size() && packageName != null) {
             logger.warn("No classes found in package " + packageName);
         }